requirements.txt
streamlit==1.41.1
torch==2.3.0
openai==1.70.0
anthropic==0.49.0
google-api-python-client==2.135.0
google-auth
google-auth-httplib2
google-auth-oauthlib
pydrive2
pillow==11.1.0
requests==2.32.3
transformers==4.48.2
streetview
ollama==0.4.7
google-generativeai
